Fix a lot of style issues.

+++ latency_rec/src/latency_log.cpp
@@ -9,7 +9,7 @@
 #include <sstream>
 
 // evaluation
-void evaluate(std::vector<long long>& lat_arr_, size_t rec_size_, size_t warmups_, const std::string& log_file_)
+void evaluate(std::vector<uint64_t> & lat_arr_, size_t rec_size_, size_t warmups_, const std::string & log_file_)
 {
   std::stringstream ss;
 
@@ -55,7 +55,7 @@ void evaluate(std::vector<long long>& lat_arr_, size_t rec_size_, size_t warmups
   }
 }
 
-void log2file(const std::vector<long long>& lat_arr_, size_t rec_size_, const std::string& log_file_)
+void log2file(const std::vector<uint64_t> & lat_arr_, size_t rec_size_, const std::string & log_file_)
 {
   if (!log_file_.empty())
   {
